Ubuntu 8.10 gcc 4.3.2: compling with -Wall, warning free except for those pesky ignored #pragmas =)
git-svn-id: https://svn.code.sf.net/p/lass/code/lass/trunk@1011 e9412239-ebb7-49aa-8f4a-f9aa784616d7

+++ lass/python/export_traits_prim.h
@@ -64,7 +64,7 @@ struct PyExportTraitsVectorPoint
 	static PyObject* build(const ObjectType& v)
 	{
 		PyObject* const tuple = PyTuple_New(dimension);
-		for (int i = 0; i < dimension; ++i)
+		for (size_t i = 0; i < dimension; ++i)
 		{
 			PyTuple_SetItem(tuple, i, pyBuildSimpleObject(v[i]));
 		}
@@ -81,7 +81,7 @@ struct PyExportTraitsVectorPoint
 		}
 		PyObject** objects = PySequence_Fast_ITEMS(tuple.get());
 		ObjectType result;
-		for (int k = 0; k < dimension; ++k)
+		for (size_t k = 0; k < dimension; ++k)
 		{
 			if (!impl::decodeObject(objects[k], result[k], k + 1))
 			{
